refactor: Clean codebase and separate Store controllers
