+++ src/interfaces/http/modules/store/controllers/index.ts
@@ -1,0 +1,14 @@
+import {StoreController} from './store.controller';
+import {StoreKiosksController} from './store-kiosks.controller';
+import {StoreIngredientsController} from './store-ingredients.controller';
+
+export const StoreContollers = [
+	StoreController,
+	StoreKiosksController,
+	StoreIngredientsController
+];
+export {
+	StoreController,
+	StoreKiosksController,
+	StoreIngredientsController
+};
